28. video bitti rezervasyon urunlerini cart ekleme listeleme islemleri yapildi

+++ product/models.py
@@ -85,7 +85,7 @@ class Room(models.Model):
     title = models.CharField(max_length=50, blank=True)
     image = models.ImageField(blank=True, upload_to='images/')
     description = models.CharField(blank=True, max_length=255)
-    price = models.IntegerField(blank=True)
+    price = models.FloatField(blank=True)
     status = models.CharField(max_length=10, choices=STATUS)
     create_at = models.DateTimeField(auto_now_add=True)
     update_at = models.DateTimeField(auto_now=True)
